+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [unreleased]
+
+### Fix
+- handle dom parser exception for invalid xml (#821)
+
+
 ## [6.16.3] (2024-10-30)
 
 ### Fix